third_party/WebKit/LayoutTests/media/video-display-aspect-ratio-expected.txt

EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '427') OK
EXPECTED (video.videoHeight == '240') OK
END OF TEST
